refactor global flag funcs to their own file
